+++ src/models/parameters.ts
@@ -145,11 +145,11 @@ export const wait: OperationParameter = {
 export const rejectDuplicates: OperationParameter = {
   parameterPath: ["options", "rejectDuplicates"],
   mapper: {
-    defaultValue: false,
+    defaultValue: "false",
     serializedName: "rejectDuplicates",
     xmlName: "rejectDuplicates",
     type: {
-      name: "Boolean"
+      name: "String"
     }
   }
 };